requirements.txt
future==0.17.1
httpx==0.22.0
lxml==4.2.5
matplotlib==3.0.0
numpy==1.15.2
pandas==0.23.4
scapy==2.4.0
scikit-datasets==0.1.15
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
sklearn-pandas==1.7.0
urllib3==1.23
xgboost==1.5.2
torch==1.10.2